rpi/requirements.txt
pydantic~=2.10.6
numpy~=2.2.3
sympy~=1.13.3
PyYAML~=6.0.2
pyserial~=3.5